Revert "HW_3"
This reverts commit 505b24baf827129479b65143df1c25e68b96bd19.
