+++ app/src/main/java/com/example/pictureoftheday/BottomNavigationDrawerFragment.kt
@@ -1,11 +1,10 @@
-package com.example.pictureoftheday.potd
+package com.example.pictureoftheday
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
-import com.example.pictureoftheday.R
 import com.example.pictureoftheday.databinding.BottomNavigationLayoutBinding
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
