+++ app/src/main/java/com/example/pictureoftheday/PictureOfTheDayAPI.kt
@@ -1,4 +1,4 @@
-package com.example.pictureoftheday.potd
+package com.example.pictureoftheday
 
 import retrofit2.Call
 import retrofit2.http.GET
