--- app/src/main/java/com/example/pictureoftheday/api/ViewPageAdapter.kt
@@ -1,32 +1,0 @@
-package com.example.pictureoftheday.api
-
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentManager
-import androidx.fragment.app.FragmentPagerAdapter
-import com.example.pictureoftheday.api.epic.EarthFragment
-import com.example.pictureoftheday.api.marsrover.MarsFragment
-
-private const val EARTH_FRAGMENT = 0
-private const val MARS_FRAGMENT = 1
-private const val WEATHER_FRAGMENT = 2
-
-class ViewPageAdapter(private val fragmentManager: FragmentManager) : FragmentPagerAdapter(fragmentManager){
-    private val fragments = arrayOf(EarthFragment(), MarsFragment(), WeatherFragment())
-
-    override fun getItem(position: Int): Fragment {
-        return when (position) {
-            0 -> fragments[EARTH_FRAGMENT]
-            1 -> fragments[MARS_FRAGMENT]
-            2 -> fragments[WEATHER_FRAGMENT]
-            else -> fragments[EARTH_FRAGMENT]
-        }
-    }
-
-    override fun getCount(): Int {
-        return fragments.size
-    }
-
-    override fun getPageTitle(position: Int): CharSequence? {
-        return null
-    }
-}
