--- app/src/main/java/com/example/pictureoftheday/api/WeatherFragment.kt
@@ -1,19 +1,0 @@
-package com.example.pictureoftheday.api
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.example.pictureoftheday.R
-
-class WeatherFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_weather, container, false)
-    }
-}
