--- app/src/main/java/com/example/pictureoftheday/api/epic/EPICData.kt
@@ -1,7 +1,0 @@
-package com.example.pictureoftheday.api.epic
-
-sealed class EPICData {
-    data class Success(val serverResponseData: EPICServerResponseData) : EPICData()
-    data class Error(val error: Throwable) : EPICData()
-    data class Loading(val progress: Int?) : EPICData()
-}
