--- app/src/main/java/com/example/pictureoftheday/api/epic/EPICServerResponseData.kt
@@ -1,22 +1,0 @@
-package com.example.pictureoftheday.api.epic
-
-import org.json.JSONArray
-
-data class EPICServerResponseJSONArray (
-    val jsonArray: JSONArray
-//    val image: String?,
-//    val date: String?,
-//    val caption: String?
-)
-
-data class EPICServerResponseData(
-    val image: String?,
-    val date: String?,
-    val caption: String?
-)
-//
-//data class EpicResponseDTO(
-//    val image: String?,
-//    val date: String?,
-//    val caption: String?
-//)
