--- app/src/main/java/com/example/pictureoftheday/utilities/RetrofitImpl.kt
@@ -1,72 +1,0 @@
-package com.example.pictureoftheday.utilities
-
-import com.example.pictureoftheday.api.epic.EPICAPI
-import com.example.pictureoftheday.api.marsrover.MarsRoverPhotosAPI
-import com.example.pictureoftheday.potd.PictureOfTheDayAPI
-import com.google.gson.GsonBuilder
-import okhttp3.Interceptor
-import okhttp3.OkHttpClient
-import okhttp3.Response
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import java.io.IOException
-
-class RetrofitImpl {
-    private val baseURL = "https://api.nasa.gov/"
-
-    fun getRetrofitImpl(): PictureOfTheDayAPI {
-        val podRetrofit = Retrofit.Builder()
-            .baseUrl(baseURL)
-            .addConverterFactory(GsonConverterFactory.create(GsonBuilder().setLenient().create()))
-            .client(createOkHttpClient(PODInterceptor()))
-            .build()
-        return podRetrofit.create(PictureOfTheDayAPI::class.java)
-    }
-
-    fun getRetrofitImplEPIC(): EPICAPI {
-        val epicRetrofit = Retrofit.Builder()
-            .baseUrl(baseURL)
-            .addConverterFactory(GsonConverterFactory.create(GsonBuilder().setLenient().create()))
-            .client(createOkHttpClient(EPICInterceptor()))
-            .build()
-        return epicRetrofit.create(EPICAPI::class.java)
-    }
-
-    fun getRetrofitImplMars() : MarsRoverPhotosAPI {
-        val marsRetrofit = Retrofit.Builder()
-            .baseUrl(baseURL)
-            .addConverterFactory(GsonConverterFactory.create(GsonBuilder().setLenient().create()))
-            .client(createOkHttpClient(MarsInterceptor()))
-            .build()
-        return marsRetrofit.create(MarsRoverPhotosAPI::class.java)
-    }
-
-    private fun createOkHttpClient(interceptor: Interceptor): OkHttpClient {
-        val httpClient = OkHttpClient.Builder()
-        httpClient.addInterceptor(interceptor)
-        httpClient.addInterceptor(HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY))
-        return httpClient.build()
-    }
-
-    inner class PODInterceptor : Interceptor {
-        @Throws(IOException::class)
-        override fun intercept(chain: Interceptor.Chain): Response {
-            return chain.proceed(chain.request())
-        }
-    }
-
-    inner class EPICInterceptor : Interceptor {
-        @Throws(IOException::class)
-        override fun intercept(chain: Interceptor.Chain): Response {
-            return chain.proceed(chain.request())
-        }
-    }
-
-    inner class MarsInterceptor : Interceptor {
-        @Throws(IOException::class)
-        override fun intercept(chain: Interceptor.Chain): Response {
-            return chain.proceed(chain.request())
-        }
-    }
-}
