--- app/src/main/java/com/example/pictureoftheday/utilities/ViewSnackbar.kt
@@ -1,11 +1,0 @@
-package com.example.pictureoftheday.utilities
-
-import android.view.View
-import com.google.android.material.snackbar.Snackbar
-
-fun View.showSnackbar(
-    message: String,
-    length: Int
-) {
-    Snackbar.make(this, message, length).show()
-}
