--- app/src/main/res/drawable-v24/bg_system.xml
@@ -1,84 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="512dp"
-    android:height="512dp"
-    android:viewportWidth="512"
-    android:viewportHeight="512">
-  <path
-      android:pathData="M257.654,266.615m-199.143,0a199.143,199.143 0,1 1,398.286 0a199.143,199.143 0,1 1,-398.286 0"
-      android:fillColor="#598889"/>
-  <path
-      android:pathData="M257.654,266.615m-199.143,0a199.143,199.143 0,1 1,398.286 0a199.143,199.143 0,1 1,-398.286 0"
-      android:fillColor="#598889"/>
-  <path
-      android:pathData="M481.13,266.855h-53.917c0,-0.081 0.003,-0.16 0.003,-0.241c0,-93.646 -75.916,-169.562 -169.562,-169.562c-61.32,0 -115.033,32.552 -144.813,81.315H59.388c-9.282,0 -16.875,7.594 -16.875,16.875c0,5.173 2.362,9.816 6.057,12.919H36.987c-12.645,0 -22.992,10.346 -22.992,22.992s10.347,22.992 22.992,22.992h51.565c-0.299,4.118 -0.46,8.274 -0.46,12.468c0,9.687 0.824,19.179 2.384,28.421H48.23c-12.645,0 -22.993,7.016 -22.993,15.59c0,8.573 10.348,15.588 22.993,15.588h50.645c24.125,64.242 86.108,109.964 158.78,109.964c47.101,0 89.711,-19.21 120.438,-50.216h70.323c9.281,0 16.876,-7.595 16.876,-16.876s-7.595,-16.875 -16.876,-16.875h-44.37c4.831,-8.246 8.981,-16.939 12.385,-25.997h39.772c12.646,0 22.992,-7.015 22.992,-15.588c0,-3.807 -2.043,-7.305 -5.417,-10.019h7.353c9.281,0 16.875,-7.595 16.875,-16.876S490.411,266.855 481.13,266.855z"
-      android:fillColor="#5AA096"/>
-  <path
-      android:pathData="M481.13,266.855h-53.917c0,-0.081 0.003,-0.16 0.003,-0.241c0,-93.646 -75.916,-169.562 -169.562,-169.562c-61.32,0 -115.033,32.552 -144.813,81.315H59.388c-9.282,0 -16.875,7.594 -16.875,16.875s7.594,16.875 16.875,16.875h37.653c-5.802,17.104 -8.949,35.433 -8.949,54.496c0,93.646 75.915,169.563 169.562,169.563c47.101,0 89.711,-19.21 120.438,-50.216h70.323c9.281,0 16.876,-7.595 16.876,-16.876s-7.595,-16.875 -16.876,-16.875h-44.37c9.232,-15.757 16.002,-33.132 19.761,-51.604h57.324c9.281,0 16.875,-7.595 16.875,-16.876S490.411,266.855 481.13,266.855z"
-      android:fillColor="#466C6F"/>
-  <path
-      android:pathData="M257.654,412.966c-80.699,0 -146.352,-65.652 -146.352,-146.352c0,-80.698 65.653,-146.352 146.352,-146.352c80.698,0 146.352,65.653 146.352,146.352C404.006,347.313 338.353,412.966 257.654,412.966zM257.654,134.263c-72.979,0 -132.352,59.373 -132.352,132.352c0,72.979 59.373,132.352 132.352,132.352c72.979,0 132.352,-59.372 132.352,-132.352C390.006,193.636 330.633,134.263 257.654,134.263z"
-      android:fillColor="#5AA096"/>
-  <path
-      android:pathData="M257.654,375.629c-60.111,0 -109.015,-48.903 -109.015,-109.015c0,-60.11 48.904,-109.014 109.015,-109.014c60.11,0 109.014,48.903 109.014,109.014C366.668,326.726 317.765,375.629 257.654,375.629zM257.654,171.601c-52.391,0 -95.015,42.623 -95.015,95.014c0,52.392 42.624,95.015 95.015,95.015c52.391,0 95.014,-42.623 95.014,-95.015C352.668,214.224 310.045,171.601 257.654,171.601z"
-      android:fillColor="#5AA096"/>
-  <path
-      android:pathData="M257.561,268.979m-66.433,0a66.433,66.433 0,1 1,132.866 0a66.433,66.433 0,1 1,-132.866 0"
-      android:fillColor="#2F485E"/>
-  <path
-      android:pathData="M151.381,391.946m-19.773,0a19.773,19.773 0,1 1,39.546 0a19.773,19.773 0,1 1,-39.546 0"
-      android:fillColor="#E0E0E0"/>
-  <path
-      android:pathData="M364.987,231.485m-19.773,0a19.773,19.773 0,1 1,39.546 0a19.773,19.773 0,1 1,-39.546 0"
-      android:fillColor="#FFFFFF"/>
-  <path
-      android:pathData="M181.956,348.487m-8.452,0a8.452,8.452 0,1 1,16.904 0a8.452,8.452 0,1 1,-16.904 0"
-      android:fillColor="#FFD800"/>
-  <path
-      android:pathData="M196.265,178.828m-8.452,0a8.452,8.452 0,1 1,16.904 0a8.452,8.452 0,1 1,-16.904 0"
-      android:fillColor="#52C9D5"/>
-  <path
-      android:pathData="M403.825,117.016m-19.773,0a19.773,19.773 0,1 1,39.546 0a19.773,19.773 0,1 1,-39.546 0"
-      android:fillColor="#66B2B2"/>
-  <path
-      android:pathData="M120.212,120.289m-19.773,0a19.773,19.773 0,1 1,39.546 0a19.773,19.773 0,1 1,-39.546 0"
-      android:fillColor="#66B2B2"/>
-  <path
-      android:pathData="M324.296,108.473m-19.773,0a19.773,19.773 0,1 1,39.546 0a19.773,19.773 0,1 1,-39.546 0"
-      android:fillColor="#66B2B2"/>
-  <path
-      android:pathData="M346.296,441.489m-19.773,0a19.773,19.773 0,1 1,39.546 0a19.773,19.773 0,1 1,-39.546 0"
-      android:fillColor="#66B2B2"/>
-  <path
-      android:pathData="M417.111,395.012m-19.772,0a19.772,19.772 0,1 1,39.544 0a19.772,19.772 0,1 1,-39.544 0"
-      android:fillColor="#66B2B2"/>
-  <path
-      android:pathData="M49.09,436.383m-8.452,0a8.452,8.452 0,1 1,16.904 0a8.452,8.452 0,1 1,-16.904 0"
-      android:fillColor="#66B2B2"/>
-  <path
-      android:pathData="M301.535,380.171m-8.452,0a8.452,8.452 0,1 1,16.904 0a8.452,8.452 0,1 1,-16.904 0"
-      android:fillColor="#66B2B2"/>
-  <path
-      android:pathData="M57.267,117.506m-8.452,0a8.452,8.452 0,1 1,16.904 0a8.452,8.452 0,1 1,-16.904 0"
-      android:fillColor="#66B2B2"/>
-  <path
-      android:pathData="M465.062,145.101m-8.452,0a8.452,8.452 0,1 1,16.904 0a8.452,8.452 0,1 1,-16.904 0"
-      android:fillColor="#66B2B2"/>
-  <path
-      android:pathData="M338.329,272.856m-8.452,0a8.452,8.452 0,1 1,16.904 0a8.452,8.452 0,1 1,-16.904 0"
-      android:fillColor="#66B2B2"/>
-  <path
-      android:pathData="M257.654,266.615m-29.053,0a29.053,29.053 0,1 1,58.106 0a29.053,29.053 0,1 1,-58.106 0"
-      android:fillColor="#FF7143"/>
-  <path
-      android:pathData="M258.046,237.564c-0.132,-0.002 -0.263,0.005 -0.395,0.005l-0.784,58.08c0.132,0.004 0.262,0.014 0.395,0.016c16.044,0.216 29.226,-12.614 29.442,-28.658S274.09,237.781 258.046,237.564z"
-      android:fillColor="#E5603C"/>
-  <path
-      android:pathData="M416.202,179.242l5.193,10.522l11.611,1.686l-8.402,8.19l1.983,11.564l-10.385,-5.46l-10.385,5.46l1.984,-11.564l-8.402,-8.19l11.611,-1.686z"
-      android:fillColor="#FFD326"/>
-  <path
-      android:pathData="M49.203,347.242l5.192,10.522l11.61,1.686l-8.401,8.19l1.983,11.564l-10.384,-5.46l-10.386,5.46l1.984,-11.564l-8.402,-8.19l11.611,-1.686z"
-      android:fillColor="#FFD326"/>
-  <path
-      android:pathData="M141.203,46.242l5.192,10.522l11.61,1.686l-8.401,8.19l1.983,11.564l-10.384,-5.46l-10.386,5.46l1.984,-11.564l-8.402,-8.19l11.611,-1.686z"
-      android:fillColor="#FFD326"/>
-</vector>
