--- app/src/main/res/drawable-v24/ic_earth.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="512"
-    android:viewportHeight="512">
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M437.019,74.981C388.668,26.629 324.38,0 256,0S123.333,26.629 74.981,74.981C26.629,123.333 0,187.62 0,256s26.629,132.667 74.981,181.019C123.333,485.371 187.62,512 256,512s132.668,-26.629 181.019,-74.981C485.371,388.667 512,324.38 512,256S485.371,123.333 437.019,74.981zM375.98,110.031l11.254,-46.752c44.813,30.612 78.449,76.426 93.423,129.958l-5.895,1.134c-8.803,1.693 -15.888,8.409 -18.046,17.112l-2.287,9.218c-0.025,0.104 -0.042,0.172 -0.209,0.231c-0.165,0.059 -0.222,0.017 -0.308,-0.047l-13.659,-10.243c-7.08,-5.31 -16.72,-6.089 -24.558,-1.984l-17.046,8.924c-8.175,4.28 -12.901,12.661 -12.333,21.873c0.568,9.211 6.287,16.948 14.93,20.193l0.808,0.303c0.093,0.035 0.162,0.06 0.204,0.228c0.042,0.166 -0.005,0.221 -0.072,0.297l-10.618,12.211c-0.107,0.122 -0.291,0.15 -0.429,0.06l-28.62,-18.466c-6.563,-4.235 -14.744,-4.873 -21.882,-1.702c-7.138,3.17 -12.154,9.663 -13.414,17.37l-2.208,13.493c-1.248,7.616 1.372,15.312 7.009,20.586l14.926,13.967c0.118,0.11 0.139,0.288 0.05,0.422l-24.326,36.875c-2.314,3.509 -3.628,7.586 -3.797,11.789l-2.111,52.793c-0.003,0.083 -0.038,0.162 -0.099,0.221l-32.395,31.753c-0.107,0.105 -0.272,0.125 -0.399,0.05l-21.006,-12.305c-0.08,-0.047 -0.137,-0.125 -0.156,-0.214c0,0 -10.191,-46.244 -10.572,-47.269l-9.265,-24.868c-3.096,-8.308 -10.802,-14.172 -19.632,-14.939l-42.402,-3.688c-0.107,-0.009 -0.203,-0.069 -0.257,-0.164l-24.122,-41.591c-0.038,-0.065 -0.051,-0.142 -0.041,-0.215l6.067,-40.337c0.013,-0.086 0.058,-0.163 0.127,-0.215l37.037,-27.945c0.094,-0.071 0.218,-0.085 0.327,-0.041l35.888,14.972c2.043,0.851 4.194,1.404 6.391,1.643l49.706,5.403c5.666,0.616 11.319,-0.873 15.944,-4.188l32.932,-23.605c8.872,-6.359 12.081,-18.209 7.63,-28.177l-3.055,-6.838c-3.545,-7.936 -11.139,-13.161 -19.816,-13.635l-92.487,-5.064c-0.05,-0.003 -0.1,-0.018 -0.144,-0.043c-0.094,-0.053 -0.144,-0.096 -0.149,-0.096v0.001c-0.045,-0.104 -0.011,-0.367 0.042,-0.44c0.014,-0.009 0.072,-0.039 0.185,-0.071l27.766,-7.484c3.673,-0.99 7.075,-2.892 9.838,-5.5l33.956,-32.043c0.051,-0.049 0.119,-0.08 0.191,-0.088l51.104,-5.263C365.623,126.591 373.692,119.535 375.98,110.031zM202.878,91.244l19.104,-35.958c2.823,-5.315 3.713,-11.489 2.503,-17.384l-2.594,-12.645c11.137,-1.64 22.523,-2.501 34.108,-2.501c11.684,0 23.167,0.875 34.395,2.541c0,0 -2.737,2.992 -3.21,3.181l-27.057,10.853c-3.866,1.552 -7.342,4.045 -10.052,7.211l-42.157,49.274c-0.573,0.67 -1.407,1.075 -2.289,1.113c-1.339,0.059 -2.166,-0.68 -2.553,-1.11C201.577,94.151 202.732,91.522 202.878,91.244zM103.424,79.718v34.139c0,9.535 -4.779,18.318 -12.777,23.506L66.15,153.259c-5.677,3.686 -10.49,8.385 -14.302,13.972l-22.234,32.58C41.375,152.438 67.641,110.732 103.424,79.718zM118.295,415.585l-3.223,26.139C78.6,413.982 50.51,375.766 35.337,331.594h14.486c1.775,0 3.413,0.899 4.369,2.389l29.275,45.614c1.32,2.048 2.89,3.914 4.688,5.552l28.479,25.975C117.874,412.251 118.5,413.924 118.295,415.585zM256,489.244c-43.759,0 -84.736,-12.123 -119.767,-33.172l4.647,-37.7c1.104,-8.977 -2.23,-17.977 -8.909,-24.064c0,0 -29.104,-26.624 -29.355,-27l-29.275,-45.613c-5.166,-8.055 -13.961,-12.857 -23.518,-12.857h-21.03c-3.948,-16.984 -6.038,-34.672 -6.038,-52.838c0,-1.962 0.026,-3.916 0.074,-5.866l47.815,-70.069c2.105,-3.095 4.756,-5.689 7.896,-7.725l24.485,-15.883c14.495,-9.398 23.154,-25.327 23.154,-42.598V62.316c22.089,-14.854 46.848,-26.023 73.389,-32.643l2.626,12.801c0.148,0.725 0.039,1.484 -0.308,2.136l-19.102,35.957c-1.194,2.245 -2.741,8.149 -2.741,8.149c-1.551,7.852 0.547,15.84 5.757,21.916c4.968,5.795 12.086,9.057 19.67,9.057c0.369,0 11.37,0.537 19.738,-9.078l42.156,-49.271c0.333,-0.39 0.762,-0.697 1.236,-0.887l27.055,-10.852c3.848,-1.543 7.309,-4.02 10.011,-7.162l10.213,-11.875c17.977,4.78 35.113,11.644 51.117,20.34c-0.264,0.604 -13.142,53.804 -13.142,53.804c-0.032,0.135 -0.148,0.236 -0.287,0.25l-51.107,5.263c-5.022,0.518 -9.809,2.71 -13.478,6.174l-33.956,32.043c-0.039,0.036 -0.086,0.064 -0.14,0.077l-27.767,7.484c-9.059,2.441 -15.688,10.032 -16.888,19.337c-1.2,9.305 3.286,18.328 11.43,22.99c3.105,1.776 6.634,2.818 10.204,3.014l92.487,5.064c0.123,0.007 0.232,0.082 0.282,0.195l3.054,6.836c0.064,0.142 0.018,0.312 -0.11,0.403l-32.932,23.605c-0.066,0.048 -0.143,0.069 -0.226,0.06l-49.796,-5.425l-35.884,-14.97c-7.539,-3.147 -16.271,-2.043 -22.793,2.876l-37.038,27.946c-4.85,3.66 -8.019,8.985 -8.922,14.994l-6.068,40.338c-0.776,5.164 0.24,10.496 2.859,15.013l24.123,41.594c3.768,6.498 10.486,10.766 17.97,11.418l42.403,3.688c0.125,0.011 0.236,0.094 0.28,0.214l9.277,24.913l9.652,44.141c1.373,6.276 5.34,11.738 10.883,14.987l21.007,12.306c3.62,2.121 7.645,3.154 11.645,3.154c5.905,0 11.759,-2.253 16.183,-6.588l32.395,-31.753c4.22,-4.136 6.673,-9.663 6.909,-15.565l2.111,-52.789c0.002,-0.061 0.02,-0.121 0.055,-0.169l24.327,-36.875c6.213,-9.422 4.741,-21.857 -3.5,-29.567l-14.927,-13.967c-0.08,-0.075 -0.117,-0.185 -0.1,-0.294l2.21,-13.494c0.018,-0.109 0.027,-0.175 0.191,-0.248c0.159,-0.073 0.217,-0.036 0.312,0.024l28.618,18.466c9.582,6.183 22.451,4.359 29.936,-4.248l10.618,-12.211c4.962,-5.704 6.816,-13.485 4.961,-20.813c-1.856,-7.328 -7.187,-13.29 -14.268,-15.948l-0.808,-0.303c-0.104,-0.039 -0.156,-0.073 -0.158,-0.073c-0.063,-0.086 -0.077,-0.344 -0.034,-0.428c0.01,-0.01 0.058,-0.05 0.155,-0.101l17.046,-8.925c0.113,-0.059 0.249,-0.048 0.35,0.028l13.659,10.243c6.21,4.658 14.283,5.883 21.595,3.279c7.311,-2.603 12.792,-8.655 14.661,-16.189l2.287,-9.217c0.03,-0.124 0.131,-0.221 0.258,-0.245l6.647,-1.278c2.323,13.177 3.541,26.729 3.541,40.561C489.244,384.612 384.611,489.244 256,489.244z" />
-</vector>
