--- app/src/main/res/drawable-v24/ic_mars.xml
@@ -1,18 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="511.99847"
-    android:viewportHeight="511">
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="m438.41,436.082c1.781,-1.809 3.535,-3.645 5.266,-5.508 44.121,-47.48 68.324,-108.918 68.324,-174.086 0,-68.219 -26.508,-132.363 -74.648,-180.676 -10.363,-10.367 -21.559,-19.773 -33.445,-28.344 -19.191,-13.84 -40.477,-24.645 -62.762,-32.547 -22.938,-8.137 -47.273,-12.844 -71.574,-14.078 -14.109,-0.719 -28.32,-0.32 -42.363,1.254 -57.488,6.395 -110.73,31.871 -152.23,73.371 -44.285,44.285 -70.328,101.945 -74.398,163.871 -1.434,21.293 -0.195,42.793 3.645,63.785 3.469,18.938 9.086,37.477 16.738,55.141 7.668,17.691 17.363,34.492 28.824,49.996 12.328,16.68 26.684,31.906 42.637,45.168 16.168,13.434 33.934,24.914 52.883,34.016 18.652,8.957 38.395,15.594 58.656,19.762 33.563,6.906 68.43,6.992 102.051,0.395 32.215,-6.324 63.117,-18.977 90.516,-37.066 15.113,-9.98 29.172,-21.551 41.883,-34.453zM89.121,89.609c10.273,-10.27 21.332,-19.465 33.027,-27.551h111.102c5.621,0 11.094,1.086 16.285,3.242 5.18,2.141 9.82,5.242 13.789,9.207 11.805,11.82 27.512,18.328 44.227,18.328h60.637c5.523,0 10,-4.477 10,-10 0,-5.523 -4.477,-10 -10,-10h-60.637c-11.367,0 -22.051,-4.426 -30.082,-12.465 -5.84,-5.844 -12.672,-10.402 -20.281,-13.547 -7.613,-3.164 -15.668,-4.766 -23.938,-4.766h-75.961c23,-10.609 47.563,-17.406 72.875,-20.164 21.941,-2.379 44.238,-1.738 66,1.941 21.023,3.559 41.648,10.113 60.898,19.266 10.59,5.035 20.859,10.781 30.539,17.406 11.262,7.707 22.043,16.191 31.816,25.73 1.23,1.195 2.473,2.387 3.684,3.602 44.43,44.551 68.898,103.719 68.898,166.648 0,13.797 -1.176,27.414 -3.48,40.734h-196.723c-5.523,0 -10,4.477 -10,10 0,5.523 4.477,10 10,10h192.34c-7.508,28.391 -20.281,55.094 -37.84,78.914h-69.449c-6.543,0 -12.91,-1.266 -18.938,-3.766 -6.043,-2.5 -11.445,-6.109 -16.063,-10.727 -13.129,-13.129 -30.586,-20.359 -49.152,-20.359h-45.617c-18.566,0 -36.023,7.23 -49.148,20.359l-16.012,16c-9.352,9.352 -21.781,14.5 -35,14.5h-68.301c-6.18,-7.023 -11.902,-14.355 -17.152,-21.961h61.531c5.523,0 10,-4.477 10,-10 0,-5.523 -4.477,-10 -10,-10h-73.91c-10.039,-18.211 -17.594,-37.66 -22.48,-57.922 -0.031,-0.164 -0.055,-0.332 -0.094,-0.496 -2.156,-8.957 -3.777,-18.066 -4.867,-27.266 -1.496,-12.68 -1.938,-25.469 -1.414,-38.227 0.148,-3.621 0.367,-7.25 0.707,-10.859 2.711,-30.27 11.23,-59.828 25.066,-86.738h114.387c19.066,0 36.996,7.426 50.484,20.914 8.52,8.523 18.496,15.188 29.645,19.809 11.133,4.613 22.895,6.953 34.969,6.953h108.445c12.648,0 22.938,10.289 22.938,22.938 0,6.125 -2.383,11.879 -6.727,16.211 -4.32,4.328 -10.078,6.715 -16.211,6.715h-163.629c-24.402,0 -47.348,9.5 -64.617,26.762l-14.211,14.223c-13.492,13.48 -31.418,20.906 -50.477,20.906h-33.219c-5.523,0 -10,4.477 -10,10 0,5.523 4.477,10 10,10h33.219c24.398,0 47.348,-9.504 64.617,-26.766l14.211,-14.219c13.488,-13.48 31.414,-20.906 50.477,-20.906h163.633c11.48,0 22.262,-4.469 30.352,-12.57 8.113,-8.102 12.582,-18.883 12.582,-30.355 0,-23.676 -19.258,-42.938 -42.934,-42.938h-108.449c-9.434,0 -18.621,-1.828 -27.313,-5.43 -8.711,-3.609 -16.5,-8.816 -23.16,-15.477 -17.266,-17.262 -40.215,-26.77 -64.625,-26.77h-102.863c9.023,-13.945 19.57,-27.043 31.594,-39.066zM395.539,446.82c-4.93,3.621 -10,7.055 -15.199,10.277h-109.074c-2.902,0 -5.266,-2.363 -5.266,-5.27 0,-1.891 0.973,-3.141 1.563,-3.73 0.57,-0.574 1.813,-1.535 3.703,-1.535h124.625c-0.117,0.086 -0.234,0.172 -0.352,0.258zM428.707,417.301c-2.953,3.172 -5.992,6.258 -9.117,9.262h-148.324c-6.758,0 -13.105,2.633 -17.855,7.406 -4.777,4.773 -7.406,11.117 -7.406,17.855 0,13.938 11.332,25.273 25.262,25.273h68.633c-22.133,8.395 -45.551,13.453 -69.289,14.93 -18.602,1.102 -37.289,0.105 -55.652,-3.09 -4.949,-0.863 -9.871,-1.883 -14.75,-3.063 -14.07,-3.414 -27.816,-8.137 -41.168,-14.16 -12.48,-5.633 -24.34,-12.27 -35.535,-19.867 -5.836,-3.977 -11.5,-8.199 -16.957,-12.684 -0.023,-0.02 -0.043,-0.035 -0.066,-0.051 -2.738,-2.254 -5.434,-4.582 -8.086,-6.969h48.527c18.559,0 36.012,-7.23 49.137,-20.355l16.012,-16.004c9.352,-9.352 21.785,-14.5 35.012,-14.5h45.617c13.223,0 25.656,5.152 35.008,14.504 6.48,6.48 14.07,11.551 22.547,15.059 8.465,3.512 17.414,5.289 26.594,5.289h52.945c-0.359,0.391 -0.723,0.777 -1.086,1.164zM428.707,417.301" />
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="m74.945,251.203h42.867c14.98,0 27.164,-12.188 27.164,-27.168v-0.277c0,-14.98 -12.188,-27.164 -27.164,-27.164h-42.867c-14.98,0 -27.164,12.184 -27.164,27.164v0.277c0,14.98 12.184,27.168 27.164,27.168zM67.781,223.758c0,-3.953 3.211,-7.164 7.164,-7.164h42.867c3.949,0 7.164,3.211 7.164,7.164v0.277c0,3.953 -3.215,7.168 -7.164,7.168h-42.867c-3.953,0 -7.164,-3.215 -7.164,-7.168zM67.781,223.758" />
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="m265.227,303.398c-1.559,-3.742 -5.285,-6.199 -9.344,-6.168 -4.09,0.031 -7.789,2.664 -9.238,6.469 -1.543,4.059 -0.184,8.824 3.266,11.457 3.262,2.484 7.801,2.723 11.305,0.594 4.148,-2.523 5.879,-7.875 4.012,-12.352zM265.227,303.398" />
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="m271.906,148.68h112.031c5.52,0 10,-4.477 10,-10 0,-5.523 -4.48,-10 -10,-10h-112.031c-5.523,0 -10,4.477 -10,10 0,5.523 4.477,10 10,10zM271.906,148.68" />
-</vector>
