--- app/src/main/res/drawable-v24/ic_system.xml
@@ -1,51 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="480.009"
-    android:viewportHeight="480.009">
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M240.267,192.009c-13.891,-0.038 -27.117,5.944 -36.261,16.4c-1.324,-0.216 -2.66,-0.35 -4,-0.4c-17.893,-0.23 -32.584,14.089 -32.813,31.981c-0.23,17.893 14.089,32.584 31.981,32.813c17.893,0.23 32.584,-14.089 32.813,-31.981c0.128,-9.93 -4.306,-19.37 -12.029,-25.613c13.719,-11.107 33.844,-8.989 44.95,4.73c11.106,13.719 8.989,33.844 -4.73,44.95c-5.707,4.62 -12.83,7.134 -20.172,7.12c-4.407,0.014 -8.767,-0.896 -12.8,-2.672l-6.4,14.672c6.053,2.653 12.592,4.015 19.2,4c26.51,0.072 48.058,-21.36 48.131,-47.869C288.209,213.63 266.777,192.082 240.267,192.009zM200.006,256.009c-8.837,0 -16,-7.163 -16,-16s7.163,-16 16,-16s16,7.163 16,16S208.843,256.009 200.006,256.009z" />
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M380.118,244.057c-24.194,-11.028 -52.748,-0.355 -63.776,23.84c-8.48,18.604 -4.284,40.535 10.464,54.696c-45.715,47.837 -121.554,49.556 -169.391,3.841c-47.837,-45.715 -49.556,-121.554 -3.841,-169.391s121.554,-49.556 169.391,-3.841c18.067,17.266 30.283,39.749 34.937,64.303l15.72,-2.992c-14.002,-73.794 -85.174,-122.265 -158.969,-108.264c-64.242,12.189 -110.717,68.371 -110.648,133.76c-0.109,75 60.603,135.888 135.603,135.997c38.274,0.055 74.794,-16.043 100.573,-44.333c24.194,11.028 52.748,0.355 63.776,-23.84C414.986,283.639 404.312,255.085 380.118,244.057zM360.006,320.009c-17.673,0 -32,-14.327 -32,-32c0,-17.673 14.327,-32 32,-32c17.673,0 32,14.327 32,32C392.006,305.682 377.679,320.009 360.006,320.009z" />
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M360.006,288.009h16v16h-16z" />
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M184.006,48.009h16v16h-16z" />
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M40.006,176.009h16v16h-16z" />
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M400.006,224.009h16v16h-16z" />
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M432.006,32.009h16v16h-16z" />
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M176.006,416.009h16v16h-16z" />
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M376.006,464.009h16v16h-16z" />
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M8.006,456.009h16v16h-16z" />
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M80.006,424.009h16v16h-16z" />
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M152.006,264.009h16v16h-16z" />
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M88.006,64.009v16c4.418,0 8,3.582 8,8s-3.582,8 -8,8s-8,-3.582 -8,-8h-16c0,13.255 10.745,24 24,24s24,-10.745 24,-24S101.261,64.009 88.006,64.009z" />
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M458.806,141.489c8.374,-7.516 13.171,-18.228 13.2,-29.48c-0.026,-22.08 -17.92,-39.974 -40,-40c-5.895,0.027 -11.71,1.375 -17.016,3.944C341.444,-2.76 224.377,-22.645 128.966,27.369C100.004,0.771 54.963,2.687 28.365,31.649C1.767,60.611 3.683,105.652 32.645,132.25c28.962,26.598 74.003,24.682 100.601,-4.28c22.339,-24.325 25.019,-60.818 6.472,-88.145c88.015,-44.261 194.632,-25.644 262.44,45.824c-14.717,16.237 -13.485,41.331 2.752,56.048c10.855,9.839 26.249,12.896 40.04,7.952c12.617,28.451 19.11,59.237 19.056,90.36c-0.141,123.653 -100.347,223.859 -224,224c-31.674,0.047 -62.99,-6.693 -91.84,-19.768c11.211,-28.772 -3.026,-61.185 -31.798,-72.396c-17.409,-6.783 -37.055,-4.422 -52.362,6.292c-49.159,-62.32 -61.551,-146.048 -32.552,-219.936l-14.896,-5.856c-31.238,79.67 -17.666,169.948 35.616,236.912c-19.177,24.214 -15.092,59.39 9.122,78.566c24.214,19.177 59.39,15.092 78.566,-9.122c0.131,-0.166 0.262,-0.333 0.392,-0.5c31.275,14.419 65.313,21.86 99.752,21.808c132.486,-0.15 239.85,-107.514 240,-240C480.058,206.032 472.829,172.438 458.806,141.489zM80.006,136.009c-30.928,0 -56,-25.072 -56,-56c0,-30.928 25.072,-56 56,-56c30.928,0 56,25.072 56,56C135.971,110.922 110.919,135.974 80.006,136.009zM432.006,136.009c-13.255,0 -24,-10.745 -24,-24s10.745,-24 24,-24s24,10.745 24,24S445.261,136.009 432.006,136.009zM96.006,464.009c-22.091,0 -40,-17.909 -40,-40c0,-22.091 17.909,-40 40,-40s40,17.909 40,40C135.98,446.089 118.086,463.983 96.006,464.009z" />
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M448.006,408.009c-17.673,0 -32,14.327 -32,32c0,17.673 14.327,32 32,32c17.673,0 32,-14.327 32,-32C480.006,422.336 465.679,408.009 448.006,408.009zM448.006,456.009c-8.837,0 -16,-7.163 -16,-16s7.163,-16 16,-16s16,7.163 16,16S456.843,456.009 448.006,456.009z" />
-</vector>
