--- app/src/main/res/drawable-v24/ic_telescope.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="512dp"
-    android:height="512dp"
-    android:viewportWidth="512"
-    android:viewportHeight="512">
-    <path
-        android:fillColor="#FFFFFF"
-        android:pathData="M456.765,254.856l-46.821,10.579a24.231,24.231 0,0 1,-5.336 0.6,24.023 24.023,0 0,1 -23.362,-18.719L356.561,138.069a24.026,24.026 0,0 1,18.121 -28.7L421.5,98.79h0a24.026,24.026 0,0 1,28.7 18.12l24.686,109.247A24.028,24.028 0,0 1,456.765 254.856ZM121.713,223.941a8,8 0,0 0,-9.567 -6.04L79.59,225.257A24.04,24.04 0,0 0,54 214.637h0L38.39,218.164a24.027,24.027 0,0 0,-18.12 28.7l10.579,46.818a23.986,23.986 0,0 0,28.7 18.121l15.6,-3.526A24.041,24.041 0,0 0,93.7 287.684l32.556,-7.356a8,8 0,0 0,6.04 -9.568ZM390.749,443.705a8,8 0,1 1,-13.5 8.589L281.877,302.421c-0.928,0.3 -1.865,0.575 -2.823,0.791a32.067,32.067 0,0 1,-7.075 0.793,31.705 31.705,0 0,1 -10.819,-1.912L158.61,452.506a8,8 0,1 1,-13.22 -9.013L247.938,293.08a31.982,31.982 0,0 1,-4.39 -6.449l-63.425,14.33a24.2,24.2 0,0 1,-5.339 0.6,24.02 24.02,0 0,1 -23.36,-18.72l-17.632,-78.032a24,24 0,0 1,18.121 -28.7l179.476,-40.554a8,8 0,0 1,9.566 6.04L365.64,250.841a8,8 0,0 1,-6.04 9.567L304,272.972a31.79,31.79 0,0 1,-8.627 20.851ZM275.528,287.605a16.019,16.019 0,1 0,-9.405 -0.733A15.87,15.87 0,0 0,275.528 287.606ZM128,96h16v16a8,8 0,0 0,16 0L160,96h16a8,8 0,0 0,0 -16L160,80L160,64a8,8 0,0 0,-16 0L144,80L128,80a8,8 0,0 0,0 16ZM424,320a24,24 0,1 0,24 24A24,24 0,0 0,424 320Z" />
-</vector>
