--- app/src/main/res/drawable-v24/swipe_indicator_passive.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="oval">
-
-    <solid android:color="@color/colorAccent40Alfa" />
-
-    <size
-        android:width="8dp"
-        android:height="8dp" />
-</shape>
