--- app/src/main/res/layout/weather_custom_tab.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".api.ApiActivity">
-
-    <androidx.appcompat.widget.AppCompatTextView
-        android:id="@+id/tabImageTV"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:drawableStart="@drawable/ic_system"
-        android:drawablePadding="@dimen/custom_drawable_padding"
-        android:gravity="center_vertical"
-        android:text="@string/weather"
-        android:textColor="@color/colorPrimaryDark"
-        android:textSize="@dimen/custom_tab_title_text_size"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
