+++ app/src/main/res/menu/menu_bottom_bar.xml
@@ -2,12 +2,6 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <item
-        android:id="@+id/app_bar_api"
-        android:icon="@drawable/ic_telescope"
-        android:title="@string/favourite"
-        app:showAsAction="ifRoom" />
-
-    <item
         android:id="@+id/app_bar_fav"
         android:icon="@drawable/ic_favourite_menu"
         android:title="@string/favourite"
