+++ app/src/main/res/values/colors.xml
@@ -18,5 +18,4 @@
     <color name="colorOnSurface">#212121</color>
     <color name="background">#FAFAFA</color>
     <color name="colorOnBackground">#212121</color>
-    <color name="colorAccent40Alfa">#662962FF</color>
 </resources>
