+++ app/src/main/res/values/dimens.xml
@@ -3,7 +3,4 @@
     <dimen name="margin_start">16dp</dimen>
     <dimen name="margin_top">16dp</dimen>
     <dimen name="margin_end">16dp</dimen>
-    <dimen name="swipeIndicatorMargins">7dp</dimen>
-    <dimen name="custom_drawable_padding">10dp</dimen>
-    <dimen name="custom_tab_title_text_size">20sp</dimen>
 </resources>
